pyproject.toml
[project]
name = "relation-tagger"
version = "0.2.26"
dependencies = [
	"imgui-bundle",
	"pydantic",
	"numpy",
	"Pillow"
]

[project.scripts]
relation-tagger="label_tool.app:main"
